+++ src/pages/individual/Flow/index.tsx
@@ -1,5 +1,4 @@
-import FormIndividual from "src/SharedComponents/FormIndividual"
-
+import FormIndividual from "C:/Xpert_Factor_Front_BO/src/pages/individual/Flow/FormIndividual"
 const IndividualFlow = () => {
   return <FormIndividual/>
 }
